pyproject.toml
[tool.poetry]
name = "nonebot-plugin-nagabus"
version = "0.1.0"
description = ""
authors = ["ssttkkl <huang.wen.long@hotmail.com>"]
readme = "README.md"
packages = [{ include = "nonebot_plugin_nagabus" }]

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.0.0rc4"
httpx = "^0.23.0"
nonebot-adapter-onebot = "^2.2.3"
nonebot-plugin-majsoul = "^0.2.0.post1"
nonebot-plugin-datastore = "^0.6.3"
typing-extensions = "^4.6.2"
pydantic = "^1.10.8"
monthdelta = "^0.9.1"
nonebot-plugin-send-anything-anywhere = "^0.2.4"

[tool.nonebot]
plugins = ["nonebot_plugin_nagabus"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
